Implement Order Types
Change existing order type from `BridgelessOrder` to `BridgelessOrder_Base`, and then define multiple order types on top of it, essentially providing optional parameters
Stop using nonces in finding the orderHash for the simplest order type -- `BridgelessOrder_Simple`.
The usage of the *now optional* `nonce` parameter, still needs to be changed to work correctly with the new order types
Implementation of logic for "OTC" orders that specify an order `executor` is still needed as well.