BUGFIX: Fix extremely confusing 'index out of bounds' bug
The 'index out of bounds' error was arising because the constructor of `Tests` wasn't getting run prior to running the tests themselves.
I've fixed this by changing the `UsersAndSubmitter` util's setup from being in the constructor to being a function, and now `Tests`calls this function inside of its `setUp`.
Fixing that bug made it all work.